VBE_RM: error executing real mode call propagation through VBE functions
return value is extended from 2B to 4B, we might want to distinguish
when there was -1 in ax and when -1 is an error

+++ c/src/lib/libbsp/i386/pc386/include/fb_vesa.h
@@ -43,8 +43,9 @@ extern "C" {
  *                          parameters which appeared with second version
  *                          of VBE.
  * @return  register ax content as defined in VBE RETURN STATUS paragraph
+ *          -1 error calling graphical bios
  */
-uint16_t VBEControllerInformation(struct VBE_VbeInfoBlock *infoBlock, uint16_t queriedVBEVersion);
+uint32_t VBEControllerInformation(struct VBE_VbeInfoBlock *infoBlock, uint16_t queriedVBEVersion);
 
 /**
  * Fills structure infoBlock with informations about selected mode in
@@ -53,8 +54,9 @@ uint16_t VBEControllerInformation(struct VBE_VbeInfoBlock *infoBlock, uint16_t q
  * @param infoBlock pointer to the struct to be filled with mode information
  * @param modeNumber detailes of this mode to be filled
  * @return  register ax content as defined in VBE RETURN STATUS paragraph
+ *          -1 error calling graphical bios
  */
-uint16_t VBEModeInformation(struct VBE_ModeInfoBlock *infoBlock, uint16_t modeNumber);
+uint32_t VBEModeInformation(struct VBE_ModeInfoBlock *infoBlock, uint16_t modeNumber);
 
 /**
  * Sets graphics mode selected. If mode has refreshRateCtrl bit set, than the
@@ -63,16 +65,18 @@ uint16_t VBEModeInformation(struct VBE_ModeInfoBlock *infoBlock, uint16_t modeNu
  * @param modeNumber number of mode to be set
  * @param infoBlock pointer to struct containing refresh rate control info
  * @return  register ax content as defined in VBE RETURN STATUS paragraph
+ *          -1 error calling graphical bios
  */
-uint16_t VBESetMode(uint16_t modeNumber, struct VBE_CRTCInfoBlock *infoBlock);
+uint32_t VBESetMode(uint16_t modeNumber, struct VBE_CRTCInfoBlock *infoBlock);
 
 /**
  * Get currently set mode number.
  *
  * @param modeNumber variable to be filled with current mode number
  * @return  register ax content as defined in VBE RETURN STATUS paragraph
+ *          -1 error calling graphical bios
  */
-uint16_t VBECurrentMode(uint16_t *modeNumber);
+uint32_t VBECurrentMode(uint16_t *modeNumber);
 
 /**
  * Gets information about display data channel implemented in the
@@ -84,8 +88,9 @@ uint16_t VBECurrentMode(uint16_t *modeNumber);
  * @param DDCLevelSupported after call contains DDC version supported and
  *                          screen blanking state during transfer
  * @return  register ax content as defined in VBE RETURN STATUS paragraph
+ *          -1 error calling graphical bios
  */
-uint16_t VBEReportDDCCapabilities(uint16_t controllerUnitNumber, uint8_t *secondsToTransferEDIDBlock, uint8_t *DDCLevelSupported);
+uint32_t VBEReportDDCCapabilities(uint16_t controllerUnitNumber, uint8_t *secondsToTransferEDIDBlock, uint8_t *DDCLevelSupported);
 
 /**
  * Reads selected EDID block from display attached to controller's interface.
@@ -94,8 +99,9 @@ uint16_t VBEReportDDCCapabilities(uint16_t controllerUnitNumber, uint8_t *second
  * @param EDIDBlockNumber block no. to be read from the display
  * @param buffer place to store block fetched from the display 
  * @return  register ax content as defined in VBE RETURN STATUS paragraph
+ *          -1 error calling graphical bios
  */
-uint16_t VBEReadEDID(uint16_t controllerUnitNumber, uint16_t EDIDBlockNumber, union edid *buffer);
+uint32_t VBEReadEDID(uint16_t controllerUnitNumber, uint16_t EDIDBlockNumber, union edid *buffer);
 
 #ifdef __cplusplus
 }
